Separate components and unit tests
