+++ src/App.tsx
@@ -1,4 +1,4 @@
-import ToDoPage from './ToDoPage';
+import ToDoPage from './components/ToDoPage';
 import './App.css';
 
 function App() {
