+++ src/components/ToDoPage.tsx
@@ -1,0 +1,117 @@
+import React, { useEffect, useReducer, useRef, useState } from 'react';
+
+import reducer, { initialState, persistTodos } from '../store/reducer';
+import {
+  setTodos,
+  updateTodoStatus,
+  createTodo,
+  toggleAllTodos,
+  deleteAllTodos,
+} from '../store/actions';
+import Service from '../service';
+import { Todo, TodoStatus } from '../models/todo';
+import { isTodoCompleted } from '../utils';
+import TodoItem from './TodoItem';
+
+type EnhanceTodoStatus = TodoStatus | 'ALL';
+
+
+const ToDoPage = () => {
+  const [{ todos }, dispatch] = useReducer(reducer, initialState);
+  const [showing, setShowing] = useState<EnhanceTodoStatus>('ALL');
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    (async () => {
+      const resp = await Service.getTodos();
+
+      dispatch(setTodos(resp || []));
+      persistTodos(todos);
+    })()
+  }, [todos]);
+
+  const onCreateTodo = async (e: React.KeyboardEvent<HTMLInputElement>) => {
+    // Check whether input is empty, do not add todo
+    if (e.key === 'Enter' && inputRef.current && inputRef.current.value) {
+      const resp = await Service.createTodo(inputRef.current.value);
+      dispatch(createTodo(resp));
+      inputRef.current.value = '';
+    }
+  }
+
+  const onUpdateTodoStatus = (e: React.ChangeEvent<HTMLInputElement>, todoId: string) => {
+    dispatch(updateTodoStatus(todoId, e.target.checked))
+  }
+
+  const onToggleAllTodo = (e: React.ChangeEvent<HTMLInputElement>) => {
+    dispatch(toggleAllTodos(e.target.checked))
+  }
+
+  const onDeleteAllTodo = () => {
+    dispatch(deleteAllTodos());
+  }
+
+  const activeTodos = todos.reduce(function (accum, todo) {
+    return isTodoCompleted(todo) ? accum : accum + 1;
+  }, 0);
+
+  const showTodos = todos.filter((todo: Todo) => {
+    switch (showing) {
+      case TodoStatus.ACTIVE:
+        return todo.status === TodoStatus.ACTIVE;
+      case TodoStatus.COMPLETED:
+        return todo.status === TodoStatus.COMPLETED;
+      default:
+        return true;
+    }
+  });
+
+  return (
+    <div className="ToDo__container">
+      <div className="Todo__creation">
+        <input
+          ref={inputRef}
+          className="Todo__input"
+          placeholder="What need to be done?"
+          onKeyDown={onCreateTodo}
+        />
+      </div>
+      <div className="ToDo__list">
+        {
+          showTodos.map((todo: Todo) => (
+            <TodoItem
+              key={todo.id}
+              todo={todo}
+              onUpdateTodoStatus={onUpdateTodoStatus}
+            />
+          ))
+        }
+      </div>
+      <div className="Todo__toolbar">
+        {todos.length > 0 ?
+          <input
+            type="checkbox"
+            checked={activeTodos === 0}
+            onChange={onToggleAllTodo}
+          /> : <div />
+        }
+        <div className="Todo__tabs">
+          <button className="Action__btn" onClick={() => setShowing('ALL')}>
+            All
+          </button>
+          <button className="Action__btn" onClick={() => setShowing(TodoStatus.ACTIVE)}>
+            Active
+          </button>
+          <button className="Action__btn" onClick={() => setShowing(TodoStatus.COMPLETED)}>
+            Completed
+          </button>
+        </div>
+        <button className="Action__btn" onClick={onDeleteAllTodo}>
+          Clear all todos
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default ToDoPage;
